remove compositor package
