--- packages/compositor/.prettierrc.js
@@ -1,14 +1,0 @@
-module.exports = {
-  singleQuote: true,
-  bracketSpacing: false,
-  useTabs: true,
-  overrides: [
-    {
-      files: ["*.yml"],
-      options: {
-        singleQuote: false,
-      },
-    },
-  ],
-  plugins: [require.resolve("prettier-plugin-organize-imports")],
-};
