--- packages/compositor/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@remotion/compositor",
-  "version": "3.3.16",
-  "description": "Image compositor written in Rust",
-  "main": "dist/index.js",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "node build.mjs && tsc -d",
-    "watch": "tsc -w",
-    "build-all": "sh cross_compile.sh && sh copy.sh",
-    "prepublishOnly": "sh cross_compile.sh && sh copy.sh"
-  },
-  "author": "Jonny Burger <jonny@remotion.dev>",
-  "license": "SEE LICENSE IN LICENSE.md",
-  "repository": {
-    "url": "https://github.com/remotion-dev/remotion"
-  },
-  "bugs": {
-    "url": "https://github.com/remotion-dev/remotion/issues"
-  },
-  "optionalDependencies": {
-    "@remotion/compositor-darwin-arm64": "3.3.11",
-    "@remotion/compositor-darwin-x64": "3.3.11",
-    "@remotion/compositor-linux-arm64-gnu": "3.3.11",
-    "@remotion/compositor-linux-x64-gnu": "3.3.11",
-    "@remotion/compositor-linux-arm64-musl": "3.3.11",
-    "@remotion/compositor-linux-x64-musl": "3.3.11",
-    "@remotion/compositor-win32-x64-msvc": "3.3.11"
-  },
-  "devDependencies": {
-    "prettier": "^2.7.1",
-    "prettier-plugin-organize-imports": "^2.3.4",
-    "typescript": "^4.7.0",
-    "eslint": "8.25.0",
-    "@jonny/eslint-config": "3.0.266"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
