--- packages/compositor/src/index.ts
@@ -1,2 +1,0 @@
-export {compose} from './compose';
-export {getExecutablePath} from './get-executable-path';
