--- packages/compositor/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../tsconfig.settings.json",
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "references": [],
-  "exclude": ["vitest.config.ts", "dist"]
-}
