+++ packages/renderer/compositor.rs
