+++ packages/renderer/copy.sh
