+++ packages/renderer/errors.rs
