+++ packages/renderer/install_platforms.sh
