+++ packages/renderer/src/compositor/get-executable-path.ts
