+++ packages/renderer/src/compositor/payload.ts
